requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0     # For making HTTP requests to the OpenAI API
python-dotenv==1.0.0  # To load environment variables
